requirements.txt
psutil
Pillow
telethon
